requirements.txt
openai-whisper==20231117
torch>=2.0.0
numpy>=1.24.3
tqdm>=4.65.0
more-itertools>=9.1.0
transformers>=4.30.0
ffmpeg-python>=0.2.0
streamlit-audiorecorder==0.0.6
